test: Mock rdf-cube-view-query in jest setup
